Move packets into a Packets folder and any classes deemed as Utilities to a Utils folder.
