+++ src/Packets/Elysia/AuxBmsElysia.cpp
