+++ src/Packets/Elysia/KeyMotorElysia.h
@@ -1,6 +1,6 @@
 #ifndef KEYMOTORELYSIA_H
 #define KEYMOTORELYSIA_H
-#include "../PropertyDefinition.h"
+#include "../../Utils/PropertyDefinition.h"
 #include <QObject>
 
 class KeyMotorElysia : public QObject
