+++ src/Packets/Elysia/MotorDetailsElysia.cpp
